move searchQuery to separate file
